12월/pjt_unbreakableheart/requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
black==22.10.0
boto3==1.26.3
botocore==1.29.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Django==3.2.13
django-appconf==1.0.5
django-bootstrap5==22.1
django-cleanup==6.0.0
django-compressor==4.1
django-extensions==3.2.1
django-imagekit==4.1.0
django-pwa==1.0.10
django-sass-processor==1.2.2
django-shortcuts==1.6
django-storages==1.13.1
django-widget-tweaks==1.4.12
executing==1.1.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
inflection==0.5.1
install==1.3.5
ipython==8.5.0
isodate==0.6.1
jedi==0.18.1
jmespath==1.0.1
libsass==0.22.0
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
oauthlib==3.2.2
parso==0.8.3
pathspec==0.10.1
pickleshare==0.7.5
pilkit==2.0
Pillow==9.3.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
protobuf==4.21.9
psycopg2-binary==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
rcssmin==1.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rjsmin==1.2.0
rsa==4.9
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
tomli==2.0.1
traitlets==5.4.0
typing_extensions==4.4.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
whitenoise==6.2.0
